arm-bsp/trusted-services: pass MM_COMM_BUFFER_SIZE to the libts build
The libts deployment expects the MM Communication buffer size to be
provided using the "MM_COMM_BUFFER_SIZE" variable.
Previously the default value which is set in the [1] was not overridden
here in the recipe because of this.

The size of the MM Communication buffers are not changed in this
commit, the page size is 0x1000.

[1] https://git.trustedfirmware.org/plugins/gitiles/TS/trusted-services.git/+/refs/tags/v1.1.0/deployments/libts/arm-linux/CMakeLists.txt#24

Signed-off-by: Bence Balogh <bence.balogh@arm.com>
Signed-off-by: Jon Mason <jon.mason@arm.com>