arm/trusted-firmware-m: Remove duplicate config in the cmake file.
The PS_NUM_ASSET is duplicated in the cmake.config and the config_tfm_target.h file
under Corstone-1000. The commit removes the one from the cmake.config and keeps the
one in the header file.

The whole rationale behind this is for the vendor to be able to override the
configuration using the cmake file.

Signed-off-by: Yogesh Wani <yogesh.wani@arm.com>
Signed-off-by: Jon Mason <jon.mason@arm.com>