CI: run optee tests in fvp-base
The optee tests were added to the fvp-base TS image in commit
eeb6441ac6a254a43b2a3ac7306e5e7430ba2e37, but they were not added to the
tests being run as part of CI.  Add that, so that they are being run.

NOTE: it adds roughly 80 minutes to the resulting job.  So, it should
not be run everywhere.

Signed-off-by: Jon Mason <jon.mason@arm.com>