arm/trusted-firmware-a: unify git repository URL
git.trustedfirmware.org is an alias for review.trustedfirmware.org. We
moved the  main recipe to use review.trustedfirmware.org last year[1]
but not all other recipes that fetch the source followed, which means
that we have to fetch TF-A multiple times.

This commit ensures that all the recipes are using the same SRC_URI, so
we just fetch TF-A once.

[1] a6a4952e ("arm/trusted-firmware-a: use correct git URL")

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Jon Mason <jon.mason@arm.com>