arm/edk2-firmware: update to edk2-stable202511
Update edk2-firmware and edk2-basetools to the 202511 stable release.

Patches from upstream were needed to get it compiling with clang.

edk2 dropped support for 32bit arm and x86, see
https://edk2.groups.io/g/devel/topic/rfc_remove_ovmf_ia32_and/114152215
Therefore, we're dropping qemuarm edk2 testing and relevant bits

Also, sbsa-ref is having a checken/egg problem with edk2 and tf-a, where
the new versions of tf-a and the older version of edk2 where having
build issues, but the newer version of edk2 has the CPU tpolotgy issue
(see https://github.com/tianocore/edk2-platforms/issues/752).  Instead
of pinning back tf-a to a legacy version and keeping edk2 at the older
version, bump both to the latest and ignore the warning.

Signed-off-by: Jon Mason <jon.mason@arm.com>